refactor: restructure the utils
