+++ src/components/Layout/TooltipContent.js
@@ -13,7 +13,7 @@ import PropTypes from 'prop-types'
 import React from 'react'
 import { useSelector } from 'react-redux'
 import { DIMENSION_TYPE_STATUS } from '../../modules/dimensionConstants.js'
-import { extractDimensionIdParts } from '../../modules/utils.js'
+import { extractDimensionIdParts } from '../../modules/dimensionIds.js'
 import { sGetMetadata } from '../../reducers/metadata.js'
 import { sGetUiInputType, sGetUiItemsByDimension } from '../../reducers/ui.js'
 import styles from './styles/Tooltip.module.css'
