+++ src/components/MainSidebar/SelectedDimensionsContext.js
@@ -11,8 +11,8 @@ import {
     DIMENSION_TYPES_PROGRAM,
     DIMENSION_TYPES_YOURS,
 } from '../../modules/dimensionConstants.js'
+import { extractDimensionIdParts } from '../../modules/dimensionIds.js'
 import { getTimeDimensions } from '../../modules/timeDimensions.js'
-import { extractDimensionIdParts } from '../../modules/utils.js'
 import { OUTPUT_TYPE_TRACKED_ENTITY } from '../../modules/visualization.js'
 import { sGetUiInputType, sGetUiLayout } from '../../reducers/ui.js'
