+++ src/modules/__tests__/dimensionIds.spec.js
@@ -1,4 +1,4 @@
-import { formatDimensionId, extractDimensionIdParts } from '../utils.js'
+import { formatDimensionId, extractDimensionIdParts } from '../dimensionIds.js'
 import {
     OUTPUT_TYPE_TRACKED_ENTITY,
     OUTPUT_TYPE_ENROLLMENT,
